refactor(#2429): move ipc to bricks/ipc per LEGO architecture
Move src/nexus/ipc/ → src/nexus/bricks/ipc/ and update all imports
across 16 source files and 23 test files. Update ruff per-file-ignores
path for driver.py. Add IPC modules with pre-existing nexus.core
imports to PREEXISTING_VIOLATIONS allowlist.
