+++ pyproject.toml
@@ -754,7 +754,7 @@ ignore = [
 "tests/e2e/self_contained/mcp/test_mcp_server_integration.py" = ["ARG", "S101", "B904"]  # raise from
 "alembic/*" = ["UP007", "E402"]  # Alembic: Union syntax, import order
 "src/nexus/backends/*.py" = ["ARG002"]  # Keep unused context params for API compatibility
-"src/nexus/ipc/driver.py" = ["ARG002"]  # Backend ABC override — params required by interface
+"src/nexus/bricks/ipc/driver.py" = ["ARG002"]  # Backend ABC override — params required by interface
 "src/nexus/core/ace/trajectory.py" = ["ARG002"]  # Keep unused permission param for API compatibility
 "src/nexus/core/cache_store.py" = ["ARG002"]  # NullCacheStore: Null Object pattern — unused args by design
 "src/nexus/core/nexus_fs.py" = ["ARG002"]  # Constructor accepts cache params for public API
