+++ src/nexus/bricks/ipc/nats_adapter.py
