+++ src/nexus/server/lifespan/ipc.py
@@ -49,7 +49,7 @@ async def startup_ipc(app: FastAPI) -> list[asyncio.Task]:
 
     # Start TTLSweeper background task
     try:
-        from nexus.ipc.sweep import TTLSweeper
+        from nexus.bricks.ipc.sweep import TTLSweeper
 
         sweeper = TTLSweeper(
             storage=ipc_storage,
