+++ tests/unit/bricks/ipc/test_cross_zone_integration.py
@@ -12,12 +12,12 @@ from typing import Any
 
 import pytest
 
-from nexus.ipc.conventions import dead_letter_path, inbox_path
-from nexus.ipc.delivery import MessageProcessor, MessageSender
-from nexus.ipc.envelope import MessageEnvelope, MessageType
-from nexus.ipc.storage.cross_zone_driver import CrossZoneStorageDriver
+from nexus.bricks.ipc.conventions import dead_letter_path, inbox_path
+from nexus.bricks.ipc.delivery import MessageProcessor, MessageSender
+from nexus.bricks.ipc.envelope import MessageEnvelope, MessageType
+from nexus.bricks.ipc.storage.cross_zone_driver import CrossZoneStorageDriver
 from nexus.services.protocols.agent_registry import AgentInfo
-from tests.unit.ipc.fakes import (
+from tests.unit.bricks.ipc.fakes import (
     InMemoryEventPublisher,
     InMemoryHotPathPublisher,
     InMemoryStorageDriver,
@@ -278,7 +278,7 @@ class TestCrossZoneIntegration:
             payload={"action": "forbidden"},
         )
 
-        from nexus.ipc.exceptions import CrossZoneDeliveryError
+        from nexus.bricks.ipc.exceptions import CrossZoneDeliveryError
 
         with pytest.raises(CrossZoneDeliveryError) as exc_info:
             await sender.send(msg)
